tests/wpt/tests/css/css-fonts/parsing/font-variant-caps-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-variant-caps with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-caps-prop">
<meta name="assert" content="font-variant-caps supports the full grammar 'normal | small-caps | all-small-caps | petite-caps | all-petite-caps | unicase | titling-caps'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-variant-caps', 'normal');
test_valid_value('font-variant-caps', 'small-caps');
test_valid_value('font-variant-caps', 'all-small-caps');
test_valid_value('font-variant-caps', 'petite-caps');
test_valid_value('font-variant-caps', 'all-petite-caps');
test_valid_value('font-variant-caps', 'unicase');
test_valid_value('font-variant-caps', 'titling-caps');
</script>
</body>
</html>
